requirements.txt
Flask==3.0.0
pymysql==1.1.0
bcrypt==4.0.1
PyJWT==2.8.0
flask-cors==4.0.0
python-dotenv==1.0.1
